refactor: update imports & dependencies on code store
