+++ frontend/src/components/AppComponent.vue
@@ -37,13 +37,13 @@
 import Block from "@/utils/block"
 import { computed, onMounted, ref, useAttrs, inject, type ComputedRef, toRefs } from "vue"
 import type { ComponentPublicInstance } from "vue"
-import { useRouter, useRoute } from "vue-router"
+import { useRouter } from "vue-router"
 import { createResource } from "frappe-ui"
-import { getComponentRoot, isHTML, getValueFromObject, setValueInObject } from "@/utils/helpers"
-import { isDynamicValue, getDynamicValue, executeUserScript, getAPIParams } from "@/utils/code"
+import { getComponentRoot, isHTML } from "@/utils/helpers"
 import { useScreenSize } from "@/utils/useScreenSize"
+import { isDynamicValue } from "@/utils/code"
 
-import useAppStore from "@/stores/appStore"
+import useCodeStore from "@/stores/codeStore"
 import { toast } from "vue-sonner"
 import type { Field } from "@/types/ComponentEvent"
 import type { DataResult } from "@/types/Studio/StudioResource"
@@ -67,7 +67,7 @@ const styles = computed(() => {
 	Object.entries(_styles).forEach(([key, value]) => {
 		if (value) {
 			if (isDynamicValue(value.toString())) {
-				_styles[key] = getDynamicValue(value.toString(), evaluationContext.value)
+				_styles[key] = codeStore.getDynamicValue(value.toString(), evaluationContext.value)
 			}
 		}
 	})
@@ -77,17 +77,14 @@ const classes = computed(() => {
 	return [attrs.class, ...props.block.getClasses()]
 })
 
-const store = useAppStore()
+const codeStore = useCodeStore()
 const repeaterContext = inject("repeaterContext", {})
 const componentContext = inject<ComputedRef | null>("componentContext", null)
 
 const evaluationContext = computed(() => {
 	return {
-		...store.variables,
-		...store.resources,
 		...repeaterContext,
 		...componentContext?.value,
-		route: store.routeObject,
 	}
 })
 
@@ -99,7 +96,7 @@ const getComponentProps = () => {
 
 	Object.entries(propValues).forEach(([propName, config]) => {
 		if (isDynamicValue(config)) {
-			propValues[propName] = getDynamicValue(config, evaluationContext.value)
+			propValues[propName] = codeStore.getDynamicValue(config, evaluationContext.value)
 		}
 	})
 	return propValues
@@ -116,7 +113,7 @@ const componentProps = computed(() => {
 // visibility
 const showComponent = computed(() => {
 	if (props.block.visibilityCondition) {
-		const value = getDynamicValue(props.block.visibilityCondition, evaluationContext.value)
+		const value = codeStore.getDynamicValue(props.block.visibilityCondition, evaluationContext.value)
 		// Handle different return types:
 		// - Boolean: return as-is
 		// - String "true"/"false": convert to boolean
@@ -137,16 +134,16 @@ const boundValue = computed({
 	get() {
 		const modelValue = props.block.componentProps.modelValue
 		if (modelValue?.$type === "variable") {
-			return getValueFromObject(store.variables, modelValue.name)
+			return codeStore.getValueFromVariable(modelValue.name)
 		} else if (isDynamicValue(modelValue)) {
-			return getDynamicValue(modelValue, evaluationContext.value)
+			return codeStore.getDynamicValue(modelValue, evaluationContext.value)
 		}
 		return modelValue
 	},
 	set(newValue) {
 		const modelValue = props.block.componentProps.modelValue
 		if (modelValue?.$type === "variable") {
-			setValueInObject(store.variables, modelValue.name, newValue)
+			codeStore.setValueInVariable(modelValue.name, newValue)
 		} else {
 			// update the prop directly if not bound to a variable
 			props.block.setProp("modelValue", newValue)
@@ -169,7 +166,7 @@ const componentEvents = computed(() => {
 				return () => {
 					const path: string[] = event.api_endpoint.split(".")
 					// get resource
-					const resource = store.resources[path[0]]
+					const resource = codeStore.resources[path[0]]
 
 					if (resource) {
 						// access and call whitelisted method
@@ -178,7 +175,7 @@ const componentEvents = computed(() => {
 						createResource({
 							url: event.api_endpoint,
 							auto: true,
-							params: getAPIParams(event.params, evaluationContext.value),
+							params: codeStore.getAPIParams(event.params, evaluationContext.value),
 							onSuccess: handleSuccess(event),
 							onError: handleError(event),
 						})
@@ -188,7 +185,7 @@ const componentEvents = computed(() => {
 				return () => {
 					const fields: Record<string, any> = {}
 					event.fields.forEach((field: Field) => {
-						fields[field.field] = getValueFromObject(store.variables, field.value)
+						fields[field.field] = codeStore.getValueFromVariable(field.value)
 					})
 					createResource({
 						url: "frappe.client.insert",
@@ -205,13 +202,7 @@ const componentEvents = computed(() => {
 				}
 			} else if (event.action === "Run Script") {
 				return () => {
-					executeUserScript(
-						event.script,
-						store.variables,
-						store.resources,
-						repeaterContext,
-						componentContext?.value,
-					)
+					codeStore.executeUserScript(event.script, repeaterContext, componentContext?.value)
 				}
 			}
 		}
@@ -225,10 +216,10 @@ const componentEvents = computed(() => {
 const handleSuccess = (event: any) => (data: DataResult) => {
 	if (event.on_success === "script") {
 		if (event.on_success_script) {
-			const variablesRefs = toRefs(store.variables)
+			const variablesRefs = toRefs(codeStore.variables)
 			const context = {
 				...variablesRefs,
-				...store.resources,
+				...codeStore.resources,
 				...repeaterContext,
 				...componentContext?.value,
 				data,
@@ -254,10 +245,10 @@ const handleSuccess = (event: any) => (data: DataResult) => {
 const handleError = (event: any) => (error: any) => {
 	if (event.on_error === "script") {
 		if (event.on_error_script) {
-			const variablesRefs = toRefs(store.variables)
+			const variablesRefs = toRefs(codeStore.variables)
 			const context = {
 				...variablesRefs,
-				...store.resources,
+				...codeStore.resources,
 				...repeaterContext,
 				...componentContext?.value,
 				error,
