+++ frontend/src/components/PropsEditor.vue
@@ -79,7 +79,6 @@ import Block from "@/utils/block"
 
 import InlineInput from "@/components/InlineInput.vue"
 import { isObjectEmpty } from "@/utils/helpers"
-import { isDynamicValue } from "@/utils/code"
 import IconButton from "@/components/IconButton.vue"
 import Link2 from "~icons/lucide/link-2"
 import Link2Off from "~icons/lucide/link-2-off"
@@ -88,6 +87,7 @@ import { useStudioCompletions } from "@/utils/useStudioCompletions"
 import type { CompletionContext } from "@codemirror/autocomplete"
 import useComponentStore from "@/stores/componentStore"
 import { getComponentProps } from "@/utils/components"
+import { isDynamicValue } from "@/utils/code"
 import useComponentEditorStore from "@/stores/componentEditorStore"
 import type { ComponentProps } from "@/types"
 import { ComponentInput } from "@/types/Studio/StudioComponent"
