+++ frontend/src/components/StudioComponentRenderer.vue
@@ -7,13 +7,16 @@ import { provide, computed } from "vue"
 import AppComponent from "@/components/AppComponent.vue"
 import Block from "@/utils/block"
 import useComponentStore from "@/stores/componentStore"
-import { getDynamicValue, isDynamicValue } from "@/utils/code"
+
+import useCodeStore from "@/stores/codeStore"
+import { isDynamicValue } from "@/utils/code"
 
 const props = defineProps<{
 	studioComponent: Block
 	evaluationContext: Object
 }>()
 const componentStore = useComponentStore()
+const codeStore = useCodeStore()
 
 const componentContext = computed(() => {
 	const context = { ...props.studioComponent.componentProps }
@@ -26,7 +29,7 @@ const componentContext = computed(() => {
 
 			Object.entries(context).forEach(([inputName, value]) => {
 				if (isDynamicValue(value)) {
-					context[inputName] = getDynamicValue(value, props.evaluationContext)
+					context[inputName] = codeStore.getDynamicValue(value, props.evaluationContext)
 				}
 			})
 		})
